change the visit_detail paradigm
then visit detail contains 2 kind of informations:
- transfers
- services

both point to care_site that contain both care unit concepts and
services concepts.

It is then possible to ask omop for MICU and MED by using span of time
within the visit detail table.

An alternative is to move the service information under transfers thanks
to the parent_id link, but it complicates the etl and has limited
advantages